requirements.txt
streamlit
yfinance
pandas
google-generativeai
requests
beautifulsoup4